feat(reports): Implements fetching reports by userId
This commit introduces a new API endpoint that allows retrieving reports based on a specific user's ID.

The endpoint enables clients to filter and access all reports associated with a given userId, enhancing the system's ability to query report data.

This functionality supports use cases such as displaying a user's submitted reports or analyzing report trends for individual users.

+++ src/main/java/com/urbaniza/authapi/repository/ReportRepository.java
@@ -11,7 +11,7 @@ import java.util.Optional;
 @Repository
 public interface ReportRepository extends JpaRepository<Report, Long> {
 
-    List<Report> findByUserId(Integer userId);
+    List<Report> findByUserId(Long userId);
 
     Optional<Report> findById(Long id);
